drakel has been frozen since 2026-06-29. PR #12 (refactor/modulith-platform,
merged 2026-08-19) brought the sprint platform onto main and re-baselined the
Flyway chain into a single V1_0__initial_schema.sql, and main has moved on
since. Take main's tree wholesale rather than reconciling two histories.

The only deliberate difference kept is the deploy surface: docker/Dockerfile
stays on drakel's multi-stage build, which compiles the jar inside the image,
because Dokploy builds from a source clone with no pre-built target/. Main's
single-stage Dockerfile expects target/*.jar to already exist and would fail
the Drakel build. .dockerignore is kept for the same reason.

The eneo/ integration is not carried over — PR #12 left it behind and that is
accepted; it remains in this branch's history if it is ever wanted back.

NOTE: the deployed Drakel database has all 47 pre-baseline migrations recorded
in flyway_schema_history. Main's single baseline has a different V1_0 checksum
and does not resolve those 46 migrations, so Flyway validation will fail until
the database is recreated. Do not deploy this without dropping the schema
first.

+++ src/main/java/se/sundsvall/caremanagement/lifecare/service/model/CalculationIncomeView.java
@@ -1,13 +1,15 @@
 package se.sundsvall.caremanagement.lifecare.service.model;
 
+import java.math.BigDecimal;
+
 /**
  * A single income row on a Lifecare normberäkning, as read for display. Dates are passed through as the raw Lifecare
  * strings.
  */
 public record CalculationIncomeView(
 	String type,
-	Double amountApplicant,
+	BigDecimal amountApplicant,
 	String applicantSearchDate,
-	Double amountCoApplicant,
+	BigDecimal amountCoApplicant,
 	String coApplicantSearchDate) {
 }
