+++ src/main/java/se/sundsvall/caremanagement/lifecare/service/model/EffectiveExpense.java
@@ -1,16 +1,18 @@
 package se.sundsvall.caremanagement.lifecare.service.model;
 
+import java.math.BigDecimal;
+
 /**
  * The effective expense for one cost at commit — the applied amount (what the citizen asked for) and the approved
- * amount (the caseworker value when set, otherwise the regelverk cap), ready to post to Lifecare FC. {@code costType}
- * is the EB cost type the FC expense-type id is resolved from against the calculation proposal; {@code bucket} routes
- * it
- * to the FC CalculationExpenses ({@code EXPENSE}) or CalculationSpecialExpenses ({@code SPECIAL_EXPENSE}) array.
+ * amount (the caseworker value when set, otherwise the regelverk cap), ready to post to Lifecare FamilyCare. {@code
+ * costType} is the financial assistance cost type the FamilyCare expense-type id is resolved from against the
+ * calculation proposal; {@code bucket} routes it to the FamilyCare CalculationExpenses ({@code EXPENSE}) or
+ * CalculationSpecialExpenses ({@code SPECIAL_EXPENSE}) array.
  */
 public record EffectiveExpense(
 	String costType,
 	String bucket,
-	Double appliedAmount,
-	Double approvedAmount,
+	BigDecimal appliedAmount,
+	BigDecimal approvedAmount,
 	String note) {
 }
