+++ src/main/java/se/sundsvall/caremanagement/types/financialassistance/api/model/LifecareCalculationPerson.java
@@ -1,10 +1,11 @@
 package se.sundsvall.caremanagement.types.financialassistance.api.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 import java.util.Objects;
 
 /**
- * A single household member on a Lifecare normberäkning — the amount the member contributes to the norm and any
+ * A single household member on a Lifecare calculation — the amount the member contributes to the norm and any
  * deviation period Lifecare applied.
  */
 @Schema(description = "A household member on a Lifecare calculation.")
@@ -17,7 +18,7 @@ public class LifecareCalculationPerson {
 	private String name;
 
 	@Schema(description = "The amount the person contributes to the norm", examples = "4500.0")
-	private Double amount;
+	private BigDecimal amount;
 
 	@Schema(description = "The start date of the deviation period, when any", examples = "2026-06-01")
 	private String deviationFromDate;
@@ -55,15 +56,15 @@ public class LifecareCalculationPerson {
 		return this;
 	}
 
-	public Double getAmount() {
+	public BigDecimal getAmount() {
 		return amount;
 	}
 
-	public void setAmount(final Double amount) {
+	public void setAmount(final BigDecimal amount) {
 		this.amount = amount;
 	}
 
-	public LifecareCalculationPerson withAmount(final Double amount) {
+	public LifecareCalculationPerson withAmount(final BigDecimal amount) {
 		this.amount = amount;
 		return this;
 	}
